requirements.txt
Flask==3.0.0
Pillow==10.0.1
google-cloud-datastore==2.18.0
google-cloud-storage==2.11.0
requests==2.31.0
google-cloud-secret-manager==2.16.4